requirements.txt
absl-py==1.4.0
ale-py==0.8.1
atari-py==0.2.6
attrs==22.2.0
boto3==1.26.118
botocore==1.29.118
cachetools==4.2.4
carla==0.9.13
certifi==2022.12.7
charset-normalizer==3.0.1
clearml==1.9.3
cloudpickle==1.6.0
cmake==3.26.3
cycler==0.11.0
decorator==4.4.2
dm-control==1.0.9
dm-env==1.6
dm-tree==0.1.8
easy-install==66.0.2
einops==0.3.0
filelock==3.12.0
furl==2.1.3
future==0.18.3
glfw==2.5.6
google-auth==2.18.1
google-auth-oauthlib==1.0.0
grpcio==1.51.3
gym==0.26.2
gym-notices==0.0.8
gym-retro==0.8.0
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
importlib-resources==5.12.0
intel-openmp==2023.1.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labmaze==1.0.6
lazy_loader==0.1
lit==16.0.5
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.4.1
mkl==2023.1.0
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.3.2
munch==2.5.0
networkx==3.0
numpy==1.21.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvidia-tensorrt==8.4.3.1
oauthlib==3.2.2
opencv-python==4.7.0.72
orderedmultidict==1.0.1
packaging==23.1
pandas==1.2.4
pathlib2==2.3.7.post1
Pillow==9.4.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
proglog==0.1.10
protobuf==3.20.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
PyJWT==2.4.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
sympy==1.12
tbb==2021.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tifffile==2023.2.27
torch==2.0.0
torch-tb-profiler==0.4.1
torchvision==0.15.1
tqdm==4.64.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
